+++ backend/common/auth.go
