+++ backend/handler/user.go
@@ -2,10 +2,18 @@ package handler
 
 import (
 	"database/sql"
+	"io/ioutil"
+	"log"
 
+	// "cloud.google.com/go/storage"
 	"github.com/aopontann/qin-todo/common"
 	"github.com/gin-gonic/gin"
 )
+
+type PutUserRequestBody struct {
+	Name  string `json:"name"`
+	Image string `json:"image"`
+}
 
 func GetUser(c *gin.Context) {
 	var (
@@ -14,7 +22,7 @@ func GetUser(c *gin.Context) {
 		email      string
 		avatar_url *sql.NullString
 	)
-	
+
 	// middlewareで認証をして成功すると、ここでユーザーIDを取得できる
 	userId := c.MustGet("userId").(string)
 
@@ -29,18 +37,36 @@ func GetUser(c *gin.Context) {
 	// この記述よくなさそうだから、他にいい方法があるか調べてみる
 	if avatar_url != nil {
 		c.JSON(200, gin.H{
-			"id":    id,
-			"name":  name,
-			"email": email,
+			"id":         id,
+			"name":       name,
+			"email":      email,
 			"avatar_url": avatar_url.String,
 		})
 		return
 	}
 	c.JSON(200, gin.H{
-		"id":    id,
-		"name":  name,
-		"email": email,
+		"id":         id,
+		"name":       name,
+		"email":      email,
 		"avatar_url": nil,
 	})
+}
 
+func PutUser(c *gin.Context) {
+	var reqb PutUserRequestBody
+	// userId := c.MustGet("userId").(string)
+	err := c.ShouldBindJSON(&reqb)
+	if err != nil {
+		c.JSON(400, gin.H{"error": err.Error()})
+		return
+	}
+
+	// client, err := storage.NewClient(c, option)
+
+	jsonFromFile, err := ioutil.ReadFile("./token.json")
+	if err != nil {
+		c.JSON(400, gin.H{"error": err.Error()})
+		return
+	}
+	log.Printf("secret=%s", string(jsonFromFile))
 }
